Update entry cache in DidFinishNavigation()
Previously we were persisting the SpeedreaderExtendedInfoHandler data to
the NavigationEntry in OnDistillComplete(). Since OnDistillComplete() is
called by the throttle, the navigation entry might not be persisted yet,
so the behavior is not defined. On backward navigations we would
sometimes persist to the previous navigation entry, causing bugs where
pages would incorrectly show up as readable. Since we wait for
DidFinishNavigation, we can safely get the last committed navgiation
entry.